Enable search
Steals a lot of code from rummager, some of which it uses, and uses rabl to render
search results as JSON. So far, I'm not convinced by rabl as it seems to be a complex
way of simply making a hash. Maybe a bit more time with it will change my mind?