Set template/views path correctly
Sinatra uses Tilt to abstract its template handling and Tilt loads the template from the file system. So
far, so good. But that means Tilt doesn't have access to the Rabl view paths and so couldn't find the
templates. Configuring the view paths at the sinatra level means everything should have access and be
able to find templates.