Add rake task to register public API in the router
We're going to be setting up a proxy to allow the public access to the content API.
That proxy needs to be registered in the router as an application and the most
obvious place to do that is from the content api app.

Changes to deployment scripts to run this task and to puppet to set up the
proxy will follow.