Remove string contain check for '*' character
The lines above each of these blocks already convert the provided
parameters to only use alphanumeric values so they're redundant.

Instead use the 'unless' idiom to fail-fast if the strings are empty.

Also updated the tests so that the Regex prevention are hard tests,
rather than soft wraps which fall back to the singular character case.