Artefact tests now use FactoryGirl, not mocks
This has several advantages, including that it makes changes
I'm about to make much easier (because the tests won't be brittle
and testing implementation).

As a side-effect it was necessary to change the code that finds an
edition for an artefact to use the panopticon_id. This is because the
Factories don't guarantee that the artefact and edition slugs will
match, so tests would fail intermittently.