Override render to always report to statsd
Rather than have independent calls to statsd for rendering in each and every
action it seems preferable to override render to do it consistently. Since
we don't (yet?) have a consistent way of working out our statsd namespace
for a given context I'm setting an @statsd_scope at the start of the action
and assuming the presence of that to record the rendering details.