Add a route to get a list of licence publications
This mirrors what we had in publisher (with a different response format) but
it feels like we may want to revisit and make it more generic, such as a
method for getting a (paginated) list of all items of a given format with
an ID within a supplied set.

My focus here is on getting the licence finder moved over to use the content
api so I'm going to leave that decision for now.