Update curated lists to be a sort ordering hint.
Instead of using the list of artefacts as the list of items to display,
they now list the top items to display in the section, with the
remaining items listed at the end. This means adding a new item to a
section doesn't require an update to the relevant curated list.

Also, this fixes an issue where an item that has been removed from a
section, or indeed has been archived, can still appear in that section's
curated list.