Fix inconsistency of the fake pagination class.
The fake class was using a method `total_pages`, rather than the `pages`
method used in the real class. The tests now test for the same methods
as are used in the `paginated.rabl` template.

As it turns out, the reason this didn't break any of the tests or views
is that RABL will fail silently if the object it is using doesn't
respond to the named attribute.

See <https://github.com/nesquena/rabl/blob/v0.6.14/lib/rabl/builder.rb#L67>