Change /tags/<type>.json to use a plural.
For example, instead of /tags/section.json, we have /tags/sections.json.

This is more consistent with the `/tags.json` URL, since it indicates a
collection response. It does make the URL harder to build manually, but we
should have a list of tag types and their URLs anyway to avoid manual
URL-building.

To avoid calling ActiveSupport's inflection methods repeatedly, there's also
a TagTypes class, which precomputes singular and plural names and acts as a
lookup table.