Increase number of unicorn processes
This should reduce the frequency of requests timing out in apps using the
content API (eg frontend, smart answers).

Our hypothesis is that the content API is simply too busy, so some requests are
queued for several seconds before the app is available. The logs for timed out
requests seem to agree with this (long nginx request duration, short app
request duration).