Remove the custom RABL error views.
There's so little information in them as they stand that it's easy to
reproduce them in straight Ruby.

As part of this, make the error response format more consistent: we were
previously using the "status" field both for short status names and for
longer-form error messages. Now we consistently have "status" for the
short status and "status_message" for longer errors.