De-RABLify the artefact list.
As part of this, I had to add an extra (optional) parameter to the URLHelper
class to supply a Plek instance to look up apps. In development, we link to
the individual apps' rendering pages, so we don't have to spool up a router on
development machines, so we need to look up the URL for an artefact's
rendering app.