Don't assume all tags have web_urls
Only tags in the 'section' tag type actually have public-facing webpages
on GOV.UK. Other tag types, such as 'legacy_sources' don't have these,
yet the Content API returns a similar looking URL pointing to a browse
page which does not exist.

This changes the URLHelper's with_tag_web_url method to check the tag
type before returning a URL, and return nil if it isn't a section tag.

Returning nil feels like the right behaviour because there isn't a
public web_url which exists for this entity. This change is unlikely to
break clients as there are currently very few tags which would be
affected by this change.