Make unencoded slashes work when requesting a tag
Right now, using an unencoded slash character in a URL for a tag ID 
returns a 404 for the tag, because Sinatra can't correctly match the 
route. Using %2F works, but this gets clobbered by Nginx when you make 
a request through www.gov.uk.

This changes the route to use a wildcard match, handling the parameters 
as block arguments instead of using the params hash.