Allow tagged artefacts to be grouped by format type
Add a new presenter, GroupedResultSetPresenter, which will group the
items in the provided result set by their kind/format.

Requests to /with_tag.json?group_by=format will receive a different
response built by this presenter, instead of the regular
ResultSetPresenter.

For now, the list of groupings is hard-coded in the new presenter, along
with the labels which should be returned for each group.