Fix test: use valid slug for detailed guides
The slug validation has been tightened up a little, and the
whitehall_artefact fixture by default generates a slug like
government/some-slug.

That slug is invalid for a detailed guide, so it was causing the slug
validation to fail since the validator has been tightend up.

Adjust this test to generate valid slugs for detailed guides.