Don't assert method call expectations in tag tests
We don't need to care about whether or not a given method is called once
or twice or more, we just care whether the correct values make it into
the output.

Asserting that methods have been called a certain number of times should
be reserved for methods that have side-effects.