Update tests to use the live_tag factory
As the recent changes to the Tag model mean that tags are now created in
draft by default, we need to explicitly create tags in the live state.

This updates all the existing factories in Content API for tags to use
the new 'live_tag' factory. Existing tags which had been manually
created in the draft state have been changed to use the 'draft_tag'
factory.