Remove Manual Presenters
With moving Manuals to the content store they are no longer sent to the
Content API. This change had a knock on effect of breaking the static
seeder because after it had a page, it loaded it from the Content API.
The presenter is decided from the format which relied on certain
information being present in order to format it like a Manual or a
ManualChangeHistory page. As this wasn't sent to the Content API, this
would then 404.

This commit removes all the code differing Manuals and Manual Change
History from a normal Artefact as all the information the seeder needs
is present in this.