Don't protect against path traversal attacks
Rack::Protection’s path traversal protection converts encoded slashes
to normal slashes before passing to the app. This is not desired as
clients of the Content API may be requesting tags with slashes in the
name, so a path with `%2F` in is actually valid.

This has only recently broken because `rack-protection` v1.5.1 changed
the regex to be case-insensitive, it was previously only looking for
`%2f` whereas our clients were passing `%2F`.

The endpoint in question is actually deprecated, the new endpoint uses
a splat url parameter to work around this. We can’t easily do that here
as the deprecated endpoint would override the correct one.