Register the `/api` prefix route via publishing api
we've introduced a new special_route format[1] to allow us to register
unusual routes which don't have a real format associated with them. This
supersedes the previous approach of registering such routes via the
router. The advantage of registering with publishing api is that the
routes will also be reserved with url-arbiter which should eliminate the
risk of clashes.

This adds a rake task to register the `/api` prefix route via that
mechanism, and removes the old method of registering directly with the
router.

The intention is to run this rake task on deploy (as we do for the route
registration at present).

[1] alphagov/govuk-content-schemas#106