Rescue and log more Licensify errors
In all environments other than current production and current staging,
Licensify requests will abort with various error like SocketError due
to not being able to resolve the licensify hostname, or
Errno::ENETUNREACH due to the application being in a different vDC that
we can’t currently route to.

In these cases, we should log the error to Errbit, but return a
friendly error to frontend allowing it to still render the rest of the
licence page.

A future improvement to this would be to attempt to distinguish between
different types of error and indicate to the user whether the licence
isn’t available to apply for online just temporarily, or if it’s not
been added to licensify at all yet.