Fix licence application test.
The original tests were stubbing a single Edition.where clause to return a
non-persisted document. However, due to a change in how Mongoid calls scopes on
related collections, this now fails when the application tries to retrieve
the live documents.

It's not clear why the stubbing was being done in the first place, although
this goes back to the original implementation of the code. Replacing this with
an actual persisted document, and dropping the stubbing altogether, works fine.