Archive the local authorities API
From our server logs it looks like nothing is using this - there have been no
requests for it in the last week. We are going to stop importing local
authorities into the shared database soon and so this data will no longer be
updated after that, so we might as well archive it now.

Also stop linking to the API from local authorities embedded in local
transactions.