Stop returning local interactions in api responses
We now store local interaction links in [Local Links
Manager](https://github.com/alphagov/local-links-manager) and use its API
to fetch them in Frontend so we don't need to return them in here any
more.  We leave the artefact+snac endpoint as it may still be useful
to retrieve and artefact with some local authority information (for
licences perhaps?).